# backend/currentopenapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CampusMind API",
        "description": "Smart Campus AI Knowledge Base",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/upload/file": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "Upload File",
                "operationId": "upload_file_api_upload_file_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_file_api_upload_file_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/files": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "List Files",
                "description": "List all institute-visible files with optional filters.",
                "operationId": "list_files_api_files_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by academic year (1-4)",
                            "title": "Year"
                        },
                        "description": "Filter by academic year (1-4)"
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by branch (e.g. COMP, AI&DS)",
                            "title": "Branch"
                        },
                        "description": "Filter by branch (e.g. COMP, AI&DS)"
                    },
                    {
                        "name": "subject",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by subject name",
                            "title": "Subject"
                        },
                        "description": "Filter by subject name"
                    },
                    {
                        "name": "doc_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by doc type (lecture|notes|pyq|lab|reference)",
                            "title": "Doc Type"
                        },
                        "description": "Filter by doc type (lecture|notes|pyq|lab|reference)"
                    },
                    {
                        "name": "file_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by file type (pdf|image|video)",
                            "title": "File Type"
                        },
                        "description": "Filter by file type (pdf|image|video)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/files/{file_id}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Get File",
                "description": "Return metadata for a single file by file_id.",
                "operationId": "get_file_api_files__file_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_file_api_upload_file_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    },
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    },
                    "branch": {
                        "type": "string",
                        "title": "Branch"
                    },
                    "subject": {
                        "type": "string",
                        "title": "Subject"
                    },
                    "unit": {
                        "type": "integer",
                        "title": "Unit"
                    },
                    "doc_type": {
                        "type": "string",
                        "title": "Doc Type"
                    }
                },
                "type": "object",
                "required": [
                    "file",
                    "year",
                    "branch",
                    "subject",
                    "doc_type"
                ],
                "title": "Body_upload_file_api_upload_file_post"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/auth/login"
                    }
                }
            }
        }
    }
}